+++ src/data/configs.js
@@ -3,7 +3,6 @@ const configs = {
     'hideOnScroll': true,
     'convertMetrics': true,
     'addOpenLinks': true,
-    'addFontFormatButtons': false,
     'addPasteButton': true,
     'convertCurrencies': true,
     'convertTime': true,